Fix overflow in autoregressive data loader

+++ lamp-data/src/main/scala/lamp/data/languagemodel/package.scala
@@ -146,7 +146,7 @@ package object languagemodel {
             )
             .toLongArray
             .toVector
-            .map(_.toInt)
+            .map(_.toLong)
 
           val tokens = STen.stack(
             starts.map(i =>
